# openapi.yaml
openapi: 3.0.3
info:
  title: Eskisini Ver Yenisini Al API
  description: |
    Trade-in platformu API'si. Kullanıcıların eski ürünlerini değerlendirip
    yeni ürünlerle takas etmelerine olanak sağlar.

    ## Domain Modeli

    ### CatalogProduct vs TradeInItem
    - **CatalogProduct**: Satışa sunulan yeni/yenilenmiş ürünler (stok, fiyat, garanti)
    - **TradeInItem**: Kullanıcının trade-in için gönderdiği eski cihaz (değerleme, durum, IMEI)

    ### Naming Conventions

    | Domain | Entity | Endpoint | Tag |
    |--------|--------|----------|-----|
    | Ürün Kataloğu | `CatalogProduct` | `/catalog/products` | Catalog |
    | Trade-In | `TradeIn` | `/trade-ins` | TradeIn |
    | Değerleme | `Valuation` | `/trade-ins/valuations` | TradeIn |
    | Sipariş | `Order` | `/orders` | Orders |
    | Kredi | `Credit` | `/credits` | Users |

    ### Terminoloji Standardı
    - **Credits** (not Wallet): Kullanıcı bakiyesi için "Credits" terminolojisi kullanılır
    - **TradeIn** (not Submission): Ana domain ismi "TradeIn", detay entity "TradeIn" olarak kalır
    - **Error**: Tüm hata response'ları `ErrorResponse` şemasını kullanır

    ### State Machine Standardı
    Tüm status alanları SCREAMING_CASE kullanır ve tutarlı geçişlere sahiptir.

    ### Settlement Policy (MVP)
    - **CREDIT**: Trade-in değeri kullanıcı hesabına kredi olarak eklenir
    - **ORDER_DISCOUNT**: Trade-in değeri siparişten düşülür
...

  version: 1.1.0
  contact:
    name: API Destek
...
  license:
...
servers:
...
paths:
...
  /auth/register:
    post:
      tags:
        - Auth
      summary: Yeni kullanıcı kaydı
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Kayıt başarılı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email zaten kayıtlı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags:
        - Auth
      summary: Kullanıcı girişi
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Giriş başarılı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags:
        - Auth
      summary: Token yenileme
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token yenilendi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags:
        - Auth
      summary: Çıkış yap
      operationId: logoutUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Çıkış başarılı
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/forgot-password:
    post:
      tags:
        - Auth
      summary: Şifre sıfırlama isteği
      operationId: forgotPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Şifre sıfırlama emaili gönderildi
        '404':
          $ref: '#/components/responses/NotFound'
  /auth/reset-password:
    post:
      tags:
        - Auth
      summary: Şifre sıfırlama
      operationId: resetPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - password
              properties:
                token:
                  type: string
                password:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Şifre başarıyla sıfırlandı
        '400':
          $ref: '#/components/responses/BadRequest'

  # ============================================
  # USERS - Kullanıcı Yönetimi
  # ============================================
...
  /catalog/products:
    get:
      tags:
        - Catalog
      summary: Katalog ürünlerini listele
      description: Satışa sunulan yeni ve yenilenmiş ürünleri listeler
      operationId: listCatalogProducts
      security: []
      parameters:
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
        - name: brandId
          in: query
          schema:
            type: string
            format: uuid
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: condition
          in: query
          schema:
            $ref: '#/components/schemas/CatalogProductCondition'
        - name: search
          in: query
          description: Arama terimi
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [PRICE_ASC, PRICE_DESC, NEWEST, POPULAR]
            default: NEWEST
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Ürün listesi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogProductListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /catalog/products/{productId}:
    get:
      tags:
        - Catalog
      summary: Katalog ürün detayı
      operationId: getCatalogProduct
      security: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ürün detayı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogProduct'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================
  # TRADE-IN - Eski Ürün Değerleme ve Geri Alım
  # ============================================
...
  /orders:
    get:
      tags:
        - Orders
      summary: Kullanıcının siparişlerini listele
      operationId: listOrders
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Sipariş listesi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Orders
      summary: Yeni sipariş oluştur
      operationId: createOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Sipariş oluşturuldu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /orders/{orderId}:
    get:
      tags:
        - Orders
      summary: Sipariş detayı
      operationId: getOrder
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Sipariş detayı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /orders/{orderId}/cancel:
    post:
      tags:
        - Orders
      summary: Siparişi iptal et
      operationId: cancelOrder
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: İptal nedeni
      responses:
        '200':
          description: Sipariş iptal edildi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Sipariş iptal edilemez (SHIPPED veya sonrası)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================
  # PAYMENTS - Ödeme İşlemleri
  # ============================================
...
  /notifications:
    get:
      tags:
        - Notifications
      summary: Bildirimleri listele
      operationId: listNotifications
      security:
        - bearerAuth: []
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
            default: false
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        '200':
          description: Bildirim listesi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /notifications/{notificationId}/read:
    post:
      tags:
        - Notifications
      summary: Bildirimi okundu olarak işaretle
      operationId: markNotificationAsRead
      security:
        - bearerAuth: []
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Bildirim okundu olarak işaretlendi
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /notifications/read-all:
    post:
      tags:
        - Notifications
      summary: Tüm bildirimleri okundu olarak işaretle
      operationId: markAllNotificationsAsRead
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Tüm bildirimler okundu olarak işaretlendi
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================
  # ADMIN - Internal Endpoints (NOT PUBLIC API)
  # ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Geçersiz istek
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Kimlik doğrulama gerekli
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Kaynak bulunamadı
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    OrderStatus:
      type: string
      enum:
        - PENDING        # Sipariş oluşturuldu, ödeme bekleniyor
        - PAID           # Ödeme alındı
        - PROCESSING     # Hazırlanıyor
        - SHIPPED        # Kargoya verildi
        - DELIVERED      # Teslim edildi
        - CANCELLED      # İptal edildi
        - RETURN_REQUESTED  # İade talebi var
        - RETURNED       # İade edildi

    # Payment Status: Ödeme durumu
    PaymentStatus:
      type: string
      enum:
        - PENDING        # Ödeme bekleniyor
        - PROCESSING     # İşleniyor
        - COMPLETED      # Başarılı
        - FAILED         # Başarısız
        - REFUNDED       # İade edildi
        - PARTIALLY_REFUNDED  # Kısmi iade

    # Shipment Status: Kargo durumu
...
    CatalogProductCondition:
      type: string
      enum:
        - NEW            # Sıfır ürün
        - REFURBISHED    # Yenilenmiş
        - LIKE_NEW       # Sıfır gibi
      description: Satıştaki ürünün durumu

    # Trade-In Item Condition (kullanıcının eski ürünü)
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - firstName
        - lastName
        - phone
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
          pattern: '^(\+90|0)?[0-9]{10}$'
        acceptTerms:
          type: boolean
          default: true
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          description: Token geçerlilik süresi (saniye)
        user:
          $ref: '#/components/schemas/User'

    # ============================================
    # USER SCHEMAS
    # ============================================
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        avatar:
          type: string
          format: uri
        emailVerified:
          type: boolean
        phoneVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Address:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: Ev
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        addressLine1:
          type: string
        addressLine2:
          type: string
        city:
          type: string
        district:
          type: string
        postalCode:
          type: string
        country:
          type: string
          default: TR
        isDefault:
          type: boolean
...
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        image:
          type: string
          format: uri
        parentId:
          type: string
          format: uuid
        children:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        tradeInSupported:
          type: boolean
          description: Bu kategoride trade-in destekleniyor mu

    # ============================================
    # BRAND & MODEL SCHEMAS
    # ============================================
...
    CatalogProduct:
      type: object
      description: Satışa sunulan yeni veya yenilenmiş ürün
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
          description: Stok kodu
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        brandId:
          type: string
          format: uuid
        brandName:
          type: string
        modelId:
          type: string
          format: uuid
        modelName:
          type: string
        categoryId:
          type: string
          format: uuid
        category:
          $ref: '#/components/schemas/Category'
        images:
          type: array
          items:
            type: string
            format: uri
        price:
          type: number
          description: Satış fiyatı (TL)
        originalPrice:
          type: number
          description: Liste fiyatı (TL)
        discountPercentage:
          type: number
          description: İndirim yüzdesi
        condition:
          $ref: '#/components/schemas/CatalogProductCondition'
        stock:
          type: integer
          minimum: 0
        specifications:
          type: object
          additionalProperties:
            type: string
        warranty:
          type: object
          properties:
            months:
              type: integer
            description:
              type: string
        tradeInEligible:
          type: boolean
          description: Bu ürün trade-in ile alınabilir mi
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CatalogProductListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CatalogProduct'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # ============================================
    # TRADE-IN ITEM SCHEMAS (Kullanıcının Eski Ürünü)
    # ============================================
...
    TimelineEvent:
      type: object
      properties:
        status:
          type: string
        timestamp:
          type: string
          format: date-time
        note:
          type: string
        actor:
          type: string
          enum: [SYSTEM, USER, ADMIN]
...
    CreateOrderRequest:
      type: object
      required:
        - items
        - shippingAddressId
      properties:
        items:
          type: array
          items:
            type: object
            required:
              - productId
              - quantity
            properties:
              productId:
                type: string
                format: uuid
              quantity:
                type: integer
                minimum: 1
        shippingAddressId:
          type: string
          format: uuid
        billingAddressId:
          type: string
          format: uuid
        shippingProviderId:
          type: string
          format: uuid
        useTradeInCredit:
          type: boolean
          default: false
          description: Mevcut trade-in kredisini kullan
        tradeInId:
          type: string
          format: uuid
          description: Bekleyen trade-in'i bu siparişe uygula (ORDER_DISCOUNT)
        couponCode:
          type: string
        notes:
          type: string
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orderNumber:
          type: string
          example: "ORD-2025-000123"
        userId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/OrderStatus'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        shippingAddress:
          $ref: '#/components/schemas/Address'
        billingAddress:
          $ref: '#/components/schemas/Address'
        subtotal:
          type: number
        shippingCost:
          type: number
        couponDiscount:
          type: number
        tradeInDiscount:
          type: number
          description: Trade-in'den uygulanan indirim
        creditUsed:
          type: number
          description: Kullanılan trade-in kredisi
        tax:
          type: number
        total:
          type: number
          description: Ödenecek toplam tutar
        payment:
          $ref: '#/components/schemas/Payment'
        shipping:
          type: object
          properties:
            provider:
              type: string
            trackingNumber:
              type: string
            estimatedDelivery:
              type: string
              format: date
        appliedTradeInId:
          type: string
          format: uuid
        notes:
          type: string
        timeline:
          type: array
          items:
            $ref: '#/components/schemas/TimelineEvent'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        productName:
          type: string
        productImage:
          type: string
          format: uri
        productCondition:
          $ref: '#/components/schemas/CatalogProductCondition'
        quantity:
          type: integer
        unitPrice:
          type: number
        totalPrice:
          type: number
    OrderListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # ============================================
    # PAYMENT SCHEMAS
    # ============================================
...
    Payment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orderId:
          type: string
          format: uuid
        amount:
          type: number
        currency:
          type: string
          default: TRY
        status:
          $ref: '#/components/schemas/PaymentStatus'
        provider:
          type: string
          enum: [IYZICO, PAYTR]
        providerPaymentId:
          type: string
        installments:
          type: integer
        paidAt:
          type: string
          format: date-time
        refundedAt:
          type: string
          format: date-time
...
    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [ORDER_UPDATE, TRADE_IN_UPDATE, PAYMENT, CREDIT, PROMOTION, SYSTEM]
        title:
          type: string
        message:
          type: string
        data:
          type: object
          description: İlgili kaynak bilgisi (orderId, submissionId, vb.)
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
    NotificationListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
        unreadCount:
          type: integer
        pagination:
          $ref: '#/components/schemas/Pagination'

    # ============================================
    # COMMON SCHEMAS
    # ============================================
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        totalItems:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time

    # ============================================
    # ADMIN SCHEMAS (Internal Use Only)
    # ============================================
...